Добавил адополнительный абстрактный класс для компонента с комментариями, добавил наследование сложным компонентам
