+++ src/view/site-menu-view.js
@@ -1,5 +1,5 @@
 import {formatFirstLetterToUpperCase} from '../utils';
-import {createElement} from '../render';
+import AbstractView from './abstract-view.js';
 
 const createFilterItemTemplate = (filter, isFirst, isActive) => {
   const {name, count} = filter;
@@ -22,27 +22,15 @@ const createSiteMenuTemplate = (filterItems) => {
   </nav>`;
 };
 
-export default class SiteMenuView {
-  #element = null;
+export default class SiteMenuView extends AbstractView {
   #filters = null;
 
   constructor(filters) {
+    super();
     this.#filters = filters;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
   }
 
   get template() {
     return createSiteMenuTemplate(this.#filters);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
